+++ verb.el
@@ -40,6 +40,7 @@
 (require 'json)
 (require 'js)
 (require 'seq)
+(require 'verb-util)
 
 (defgroup verb nil
   "An HTTP client for Emacs that extends Org mode."
@@ -289,15 +290,6 @@ a call to `verb-var'."
 (defface verb-json-key '((t :inherit font-lock-doc-face))
   "Face for highlighting JSON keys.")
 
-(defface verb-log-info '((t :inherit homoglyph))
-  "Face for highlighting I entries in the log buffer.")
-
-(defface verb-log-warning '((t :inherit warning))
-  "Face for highlighting W entries in the log buffer.")
-
-(defface verb-log-error '((t :inherit error))
-  "Face for highlighting E entries in the log buffer.")
-
 (defconst verb--http-methods '("GET" "POST" "DELETE" "PUT"
                                "OPTIONS" "HEAD" "PATCH"
                                "TRACE" "CONNECT")
@@ -306,9 +298,6 @@ a call to `verb-var'."
 (defconst verb--bodyless-http-methods '("GET" "HEAD" "DELETE" "TRACE"
                                         "OPTIONS" "CONNECT")
   "List of HTTP methods which usually don't include bodies.")
-
-(defconst verb--log-buffer-name "*Verb Log*"
-  "Default name for log buffer.")
 
 (defconst verb--url-pre-defined-headers '("MIME-Version"
                                           "Connection"
@@ -325,18 +314,8 @@ will appear duplicated in the request).")
   "Keyword to use when defining request templates.
 Request templates are defined without HTTP methods, paths or hosts.")
 
-(defconst verb--log-levels '(I W E)
-  "Log levels for the log buffer.
-I = Information.
-W = Warning.
-E = Error.")
-
 (defconst verb--http-header-regexp "^\\s-*\\([[:alnum:]_-]+:\\).*$"
   "Regexp for font locking HTTP headers.")
-
-(defconst verb--http-header-parse-regexp
-  "^\\s-*\\([[:alnum:]_-]+\\)\\s-*:\\(.*\\)$"
-  "Regexp for parsing HTTP headers.")
 
 (defconst verb--metadata-prefix "verb-"
   "Prefix for Verb metadata keys in heading properties.
@@ -530,39 +509,6 @@ more details on how to use it."
          (,verb--http-header-regexp
           (1 'verb-header)))))
 
-(define-derived-mode verb-log-mode special-mode "Verb[Log]"
-  "Major mode for displaying Verb logs.
-
-Each line contains a short message representing an event that has been
-logged:
-
-  The first part of each line is a number that represents what
-  request has generated this event (the number identifies the request).
-  If an event does not correspond to any request, \"-\" is used instead.
-  If the first part of the line is empty, then the event corresponds to
-  the same request from the previous line.
-
-  The second part of each line represents the level at which this event
-  has been logged, I for information, W for warnings and E for errors.
-
-  The third part of each line is the message itself.
-
-If this buffer is killed, it will be created again when the next
-message is logged.  To turn off logging, set `verb-enable-log' to nil."
-  (font-lock-add-keywords
-   nil '(;; (request number e.g. 10)
-         ("^[[:digit:]-]+\\s-"
-          (0 'bold))
-         ;; Log level I after request number
-         ("^[[:digit:]-]*\\s-+\\(I\\)"
-          (1 'verb-log-info))
-         ;; Log level W after request number
-         ("^[[:digit:]-]*\\s-+\\(W\\)"
-          (1 'verb-log-warning))
-         ;; Log level E after request number
-         ("^[[:digit:]-]*\\s-+\\(E\\)"
-          (1 'verb-log-error)))))
-
 (defun verb--http-method-p (m)
   "Return non-nil if M is a valid HTTP method."
   (member m verb--http-methods))
@@ -680,14 +626,6 @@ KEY and VALUE must be strings.  KEY must not be the empty string."
             (verb-toggle-show-headers))))
     (setq header-line-format nil)))
 
-(defun verb-show-log ()
-  "Switch to the *Verb Log* buffer."
-  (interactive)
-  (switch-to-buffer (get-buffer-create verb--log-buffer-name)))
-
-;; Old function name (<2.11.0)
-(defalias 'verb-view-log 'verb-show-log)
-
 (defun verb-customize-group ()
   "Show the Customize menu buffer for the Verb package group."
   (interactive)
@@ -748,41 +686,6 @@ KEY and VALUE must be strings.  KEY must not be the empty string."
                       (+ (nth 1 completions) beg -1))
                 (cddr completions))))))
 
-(defun verb--log (request level &rest args)
-  "Log a message in the *Verb Log* buffer.
-REQUEST must be a number corresponding to an HTTP request made.  LEVEL
-must be a value in `verb--log-levels'.  Use the remaining ARGS to call
-`format', and then log the result in the log buffer.
-
-If `verb-enable-log' is nil, do not log anything."
-  (setq request (if request (number-to-string request) "-"))
-  (unless (member level verb--log-levels)
-    (user-error "Invalid log level: \"%s\"" level))
-  (when verb-enable-log
-    (with-current-buffer (get-buffer-create verb--log-buffer-name)
-      (unless (derived-mode-p 'verb-log-mode)
-        (verb-log-mode))
-      (let ((inhibit-read-only t)
-            (last "")
-            (line (line-number-at-pos)))
-        ;; Get last logged request number
-        (when (re-search-backward "^\\(-\\|[[:digit:]]+\\)\\s-"
-                                  nil t)
-          (setq last (match-string 1)))
-        (goto-char (point-max))
-        ;; Log new message
-        (insert (format "%s  %s  "
-                        (if (string= last request)
-                            (make-string (length request) ? )
-                          request)
-                        level)
-                (apply #'format args)
-                "\n")
-        ;; If logged messaged contained newlines, add a blank line
-        ;; to make things more readable
-        (when (> (line-number-at-pos) (1+ line))
-          (newline))))))
-
 (defun verb--ensure-org-mode ()
   "Ensure `org-mode' is enabled in the current buffer."
   (unless (derived-mode-p 'org-mode)
@@ -792,12 +695,6 @@ If `verb-enable-log' is nil, do not log anything."
   "Ensure `verb-mode' is enabled in the current buffer."
   (unless verb-mode
     (verb-mode)))
-
-(defun verb--nonempty-string (s)
-  "Return S. If S is the empty string, return nil."
-  (if (string-empty-p s)
-      nil
-    s))
 
 (defun verb-headers-get (headers name)
   "Return value for HTTP header under NAME in HEADERS.