+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## **main** (MELPA)
 - Added support for `Verb-Proxy` heading property.
 - Removed support for the `url-queue` backend.
+- Improved `*Verb Log*` buffer formatting.
 
 ## **2.16.0** - 2024-03-02 (MELPA Stable)
 - Fixed LF being used instead of CRLF in multipart boundaries.